Corregida condición

+++ conf/3_Configuraciones_Usuario.sh
@@ -35,7 +35,7 @@ function home() {
     ## Mover al directorio Backups el archivo .bashrc
     if [[ -f "$HOME/.bashrc" ]]; then  ## Si es un archivo
         crearBackup "$HOME/.bashrc" && rm "$HOME/.bashrc"
-    else [[ -h "$HOME/.bashrc" ]]; then  ## Si es un enlace
+    elif [[ -h "$HOME/.bashrc" ]]; then  ## Si es un enlace
         rm "$HOME/.bashrc"
     fi
